+++ src/hr/avrbanac/openglplayground/renderers/MasterRenderer.java
@@ -65,6 +65,10 @@ public class MasterRenderer {
         GL11.glDisable(GL11.GL_CULL_FACE);
     }
     
+    public Matrix4f getProjectionMatrix() {
+        return projectionMatrix;
+    }
+    
     public void render(List<Light> lights, Camera camera) {
         prepare();